Move getTxHashBlockExplorerURL to networks file
