+++ app/src/components/market/common/xdai_bridge_modal/bridge_transaction_state.tsx
@@ -4,8 +4,8 @@ import styled from 'styled-components'
 
 import { CONFIRMATION_COUNT, STANDARD_DECIMALS } from '../../../../common/constants'
 import theme from '../../../../theme'
-import { networkIds } from '../../../../util/networks'
-import { formatBigNumber, getTxHashBlockExplorerURL } from '../../../../util/tools'
+import { getTxHashBlockExplorerURL, networkIds } from '../../../../util/networks'
+import { formatBigNumber } from '../../../../util/tools'
 import { TransactionStep } from '../../../../util/types'
 import { ButtonRound } from '../../../button/button_round'
 import { IconArrowUp } from '../../../common/icons/IconArrowUp'
