+++ app/src/components/modal/modal_transaction/index.tsx
@@ -5,8 +5,7 @@ import styled, { withTheme } from 'styled-components'
 
 import { CONFIRMATION_COUNT } from '../../../common/constants'
 import { useConnectedWeb3Context } from '../../../hooks'
-import { getBlockExplorer } from '../../../util/networks'
-import { getTxHashBlockExplorerURL } from '../../../util/tools'
+import { getBlockExplorer, getTxHashBlockExplorerURL } from '../../../util/networks'
 import { TransactionStep } from '../../../util/types'
 import { Button } from '../../button'
 import { ButtonType } from '../../button/button_styling_types'
